Add OmniAuth developer strategy
Add the OmniAuth developer strategy when running the application in
development.

This is only enabled when the app is running in RAILS_ENV='development'
and the setting auth_provider: developer.

Having a developer OmniAuth auth strategy makes it clearer what happens
when logging in and makes testing redirects and login buttons easier
in dev.

Add developer doc