Refactor how page setting component "None of above" works for selection answer types
is_optional is a boolean attribute and so it would make sense to treat it as such in the component.

While refactoring this, I discovered an issue with the existing specs. It was checking that the whole component contained specific text which was fine for the first two rows where the text was distinct. However this become an issue when checking for "Yes/No" The tests would pass as long as "Yes" appeared in anywhere in the component. The check for "No" would pass because one rows keys is "None of the above", so because "None" contains the combination of letters "No" it would always pass.

I've fixed the issue of the tests by insuring that we give each text a smaller content to check which in this case is checking each row and the specific value for the row (i.e excludes the key)