convert answer_settings to always use symbolized keys rather than strings
ActiveRecord's JSONB fields store data as a simple hash with string keys by default. This is in line with the JSON standard. However, it's a common requirement in Ruby and Rails to use hashes with indifferent access for convenience. One way to ensure that answer_settings attribute always uses symbols for keys is to override the getter for this attribute which is what is done with draft_question model.

The read_attribute method reads the original value of the answer_settings attribute as stored in the database. The deep_symbolize_keys method is then called on this value. This method returns a new hash where you can access values using a symbol key.

The reason deep_symbolize_keys was used and not with_indifferent_access is because of the following reasons

1. with_indifferent_access: This method allows you to access the content of the hash using either symbols or string keys interchangeably. This operation has a little overhead because a new hash object is created when invoked, but it can be handy if you aren't sure what type of keys you'll be dealing with. However, it only creates a new HashWithIndifferentAccess at the top level of the hash. If you have a nested hash structure, nested hashes will remain as Hash objects.

2. deep_symbolize_keys: This method converts all keys in the hash (including nested hashes) into symbols. This operation is more efficient than the previous one, and the result is still a plain Hash, not a HashWithIndifferentAccess object, but you will only be able to use symbols to access data, not strings.

Also, keep in mind that while this ensures the hash uses symbols for accessing data in our code, the JSONB data in the database will still be stored using string keys due to the nature of JSON.