alternative approach to signing_mou
instead of a class method on mou_signature and having to pass in the user,
we could instead use the model relationships, user & mou_signature model
are both aware of eahc other. all thats needed is a scope on the mou_signature
to return records that dont have organisation. it makes it easier to read

```
user.mou_signatures.without_organisations
```
once we have those reads we can update them all with the user org.