Fix html rendering on show live form page
Makes the what_happens_next field on the live form view render its HTML
content so that form creators have a clearer idea of what their content
will look like. It now also renders this HTML with a border, so that it
can be distinguished from the main page text.