create a start_new_question route
While working on another ticket it became clear that we should have a route to point to which will setup and start a new question rather than redirecting to one specific page in the wizard. This new route would clear down previous draft question data and then redirect to the first page in the wizard. Apart from helping us setup the "session" before starting a new question, it also means that if the wizard needs to be changed and there is a new start page then we would only have to change it in one place.