Make allowing markdown headings configurable
Allows us to choose  whether the markdown rendering, validation, and
toolbar should allow headings. In some contexts, such as the upcoming
what happens next field, we want to restrict the use of headings because
the content is rendered in multiple different contexts with different
heading structures. This could cause accessibility issues if users could
add headings.