Log hash of session.id if session exists
It's difficult to trace users behaviour over multiple pages.

Individual page requests aren't linked by any values.

Guessing based on URLs and client IP is the only way to reconstruct
user's journeys through forms.

This commit adds a hash of the user's session id to the request logs.

This value is will be stable across individual page request so can be
used to connect multiple requests.

We don't log the session id directly as it would leak information that
could be used to retrieve user's data.

This commit implements the same changes already made in
alphagov/forms-runner#487, see the messages in that PR for more
discussion.