Allow contact details and submission emails to have non-gov.uk domains
There are a number of government organisations which have domain names
that do not end with .gov.uk. We now have users on our service who work
for those organisations, as we have added an allow list of domains to
the sign up configuration (see alphagov/forms-deploy#440).

We want those users to be able to create forms on our service, including
getting form submissions by email. These users should be able to use
their government email addresses for the form contact details and
submission address.

Rather than adding a duplicate of the allow list to this app, this
commit changes the `AllowedEmailDomainValidator` to allow any email
address that has the same domain as the user. This may be an imperfect
solution, in that it limits users more than a strict allow list would,
but is easier to maintain until we have a way to share the list better.