Use translation in environment name check
In all environments except production, we show a tag in the header with
the name of the current environment. The logic for displaying the tag
relies on a hard-coded reference to "production". In the govuk-frontend
v5 update in forms-runner, we forgot to update this reference and this
caused the environment tag to appear in production.

In this commit, we change the hard-coded reference into a reference to
the environment_names.production translation, which will make errors
like this much less likely in future.