spike! Try using ActiveHash to read domains from file
This worked well for reading a YAML file into a model.

I created a file in `forms-deply/config/data/domains.yml` with

```
- domain: "gov.uk"
- govuk_content_id: "fd6cee28-ac4c-4fb5-9f46-54d24a500818"
  domain: "cefas.co.uk"
- govuk_content_id: "b548a09f-8b35-4104-89f4-f1a40bf3136d"
  domain: "dwp.gov.uk"
- govuk_content_id: "d994e55c-48c9-4795-b872-58d8ec98af12"
  domain: "mod.uk"
```

However, when it came to associating the domains with organisations, I
hit a roadblock.

From the ActiveHash docs:

> NOTE: You cannot use ActiveHash objects as children of ActiveRecord
and I don't plan on adding support for that. It doesn't really make any
sense, since you'd have to hard-code your database ids in your class or
yaml files, which is a dependency inversion.

https://github.com/active-hash/active_hash?tab=readme-ov-file#referencing-activerecord-objects-from-activehash