refactor make_live_controller create actions logic
simplified the action conditional logic by removing a level of conditional logic that would either render a partial or redirect the user to another page.

This refactor makes it quicker for the action to stop processing unnecessary code if the user simply wants to cancel the step and to go back to the forms task list page.