Change super_admin Form pundit scope
Allow super_admins to potentially list all forms.

Originally I had changed the Forms policy scope to return Form.all
. This turned out to be a mistake. ActiveResource doesn't support
scopes and no changes to the existing scope is required.

We use pundits scope functionality to limit the forms a user can list
and view.

ActiveResource does not support scopes in the same way
ActiveRecord does.

ActiveResource can merge chained where clauses into single clauses but
doesn't do anything further.

In ActiveRecord, scopes are lazy. They return relation objects which
don't make DB calls until the values are needed.

ActiveResource isn't lazy and always makes API requests to fetch
objects.

When a where clause is added to the scope, it still makes the call for
all forms and then discards the result.

If Forms where ActiveRecord objects, we could
return `scope.all`. In ActiveRecord this wouldn't instantly result in a
query. As it's an ActiveResource object, it instantly makes a call to
the API to fetch all forms.

Instead we return the Form class without any other query. There doesn't
seem to be a null query object for ActiveResource so this is the best we
can do and rely on the caller to add a query.