ruby 3.3 does not inspect an object when an exception is raised
NoMethod error usess a lot of CPU calling inspect on an object that has
raised the error. In ruby 3.3 inspect has been removed and now the exception
will only contain the objects name that had the issue instead of inspecting it.

see
https://github.com/ruby/ruby/blob/v3_3_0/NEWS.md#compatibility-issues

```
Error message for NoMethodError have changed to not use the target object's #inspect for efficiency, and says "instance of ClassName" instead. [Feature #18285]

([1] * 100).nonexisting

```