refactor how list of form captions work
previously, we were using helpers but this is not the correct place for this functionality to be. If you look at the forms index page you will see there is a lot going on in the view already, trying to calculate the caption toggling between different link href depending on if forms is draft or live.

 There is a lot of logic in the view. This makes it difficult to test the functionality and all the various states. Instead we should use a service to work out what the caption, head & body for the report should be and the view calls the service.

 This commit specifically targets captions. THe next couple commits will refactor other parts of this table