add rows attribute to FormListService and helper functions
Added a `rows` attribute to FormListService where each row represents information about a specific form passed to the service. Further, helper functions were created for generating links for form names and rendering the status related to each form.

This simplifies the view drastically by removing all the logic and gives use greater control in the service to control what columns should be shown in the report for a particular user.