add 'Created by' column for super admins in form list
This update adds a new column 'Created by' to the form list which is visible only to super admin users. They will now be able to see who created each form. Tests are also updated to ensure proper functionality.

As Form is an ActiveResource and Users are an ActiveRecord we cannot rely on relationships to put this information together. For efficiency reasons, we pull out a list of unique creator ids from the list of forms and use that to query ActiveRecord once to get a list of user id and name which is used to lookup from.