allow all editors/super admins to see who created a form
Originally, only super admins could view "created by" column in the list of forms. This was because we wanted to make sure the application could handle it. If we do deploy this and things go wrong we can easily revert this commit and we can review the situation.

The reason why only editor/super admins can see this column is because trial users should not be able to see forms created by other users (even in the same organisation) and it seemed wasteful to have a column in the report which is populated with their own name for each form they create.