Use OmniAuth setup phase to change Auth0 client for E2E tests
This commit uses the [OmniAuth setup phase](https://github.com/omniauth/omniauth/wiki/Setup-Phase)
to determine which Auth0 client to use based on the auth query param.
This allows the database connection to be disabled in the "main" client
and for home realm discovery to work as intended for enterprise
connections, unblocking [this card](https://trello.com/c/UdtU9PZC).