Move manifest and related icons into public folder
govuk-frontend v5 uses a slightly different set of icons compared to v4,
so we updated the icon paths we reference in application.html.erb to
reflect this as part of #738. It also includes a manifest for these
icons. Unfortunately, fingerprinting the icons breaks the paths in the
manifest.

To get around the broken paths, this PR adds the icons and the manifest
to the public folder, so they won't be fingerprinted. We previously did
this in forms-runner (see https://github.com/alphagov/forms-runner/pull/535/commits/fa451c2a05295efdea247cf337b630df7d989195)
and forms-product-pages (see https://github.com/alphagov/forms-product-page/commit/b470cafc5f1684422fc6b2c6df32559ebfefd263)
as part of the govuk-frontend v5 update. We should've done it in this
repo at the same time, but this was missed.

At some point we should find a way to either generate the manifest with
references to fingerprinted assets, or automatically copy the manifest
and its dependencies to `public`.