Refactor group specs to use factories
When setting up the environment for specs we now create Group records
using `factory_bot` instead of with the usual Rails/ActiveRecord
methods. The advantage of doing this is when we add/change attributes on
the model we need to do less work to bring the specs in line.

This change is also consistent with what we do in other specs in this
repo; the specs here didn't initially use factories because they were
generated with the Rails scaffold generator.

This change is particularly useful for the following commit, where we
require groups to belong to an organisation. For test setup we need
groups to be associated with an organisation, but in requests we
don't want the organisation to be in the parameters/valid attributes.