Fix `User#role_changed_to_editor?`
When a user has been created with an editor role the
`role_changed_to_editor?` method raised the exception:

```
undefined method `role' for nil:NilClass
```

This was happening because the reification of the last version was
`nil`.

I found this when playing around in local development with the
`developer` auth provider and signing in as one of the users from the
database seed. Currently in production users won't be created with the
editor role, but that might change, so I fixed it. Plus it's nice not to
have errors in development :sweat_smile:.