Rename _mou_version_1 to current
Originally this PR added a new version of the MOU and updated the views
to render it instead of the old MOU. This made it hard to see changes to
the text as no diff is available as the file is new.

This commit changes how old MOUs are stored and should make it easier to
review MOU changes.

Instead of creating a new file with the new MOU text, the current MOU
will always be called current. Archived versions of the MOU will be
numbered.

This allows us to review diffs against the current MOU text to highlight
changes and store previous versions of the MOU.

It also makes it easier to show that the contents of the MOU have not
been changed when they are archived.