Add favicon.ico to server root as symlink
We're seeing a lot of `404`s in our server logs from clients trying to
reach `/favicon.ico`. This commit hopefully fixes this by making it so
we can serve the favicon from that path.

We already have a static copy of the favicon.ico file in `public/icons`
to make the paths in the manifest.json file work (see commit b470caf),
so to avoid the need to copy the favicon twice this commit just adds a
symbolic link. If we find a way to generate the manifest file in future
(removing the need for the `public/icons` folder) we'll need to take a
different approach for this new file.