Add join model for groups and forms
We would like groups to have zero or more forms.

Ideally we would do this without adding data to the form record in the
forms-api database about data in the forms-admin database. This leads us
to using a join table, so we can associate groups with forms but also do
the reverse lookup. Normally to use a join table in Rails we use the
`has_and_belongs_to_many` association [[1]]. However, models in
ActiveResource cannot be associated with models in ActiveRecord [[2]].
This means we can't use any of the usual methods to maintain the
association.

Instead, this commit takes the approach of having an explicit join
model. It's clunky, but is the best approach I've tried so far. It
doesn't rely too much on understanding Rails magic, and it doesn't
require writing any SQL.

Probably the proper fix for this issue is keeping the form records and
the group records in the same database, but that is outside the scope of
this ticket.

[1]: https://guides.rubyonrails.org/association_basics.html#the-has-and-belongs-to-many-association
[2]: https://github.com/rails/activeresource/issues/292