Rework payment link page content
Since our initial implementation of this page, the MVP content has
changed. This commit updates the page to match those changes. The main
changes are:

- a small update to the order of the page
- a new 'Creating your payment link in GOV.UK Pay' section which covers
the specifics of setting up a GOV.UK Pay link to receive a reference
from Forms.

This commit also adds a view test to give us more confidence that this
page is displaying correctly.