Restrict access to Groups to members
Restrict access to list, view, modify and delete groups to only users
who are members of that group.

Except for super_admins, who are able to access all groups.

This commit changes the group policy. It adds tests for the existing
policy and the new scope and restrictions.

The policy scope delegates to a scope on the model because that seems a
better place to implement and test it.

To support this work the group controller has been modified so that when a
user creates a group, they are added to it.

The controller spec has also been updated because users will no longer
be able to see all groups within their organisation, only those they are
a member of.