Introduce way of synchronizing MailChimp lists with users
We have a database table full of user information, and we have some MailChimp
lists that we'd like to contain all the same email addresses.

We're happy for the synchronization to happen out-of-band with regular user
management activities, so we've built it as a Rake task that will be run on a
schedule. The scheduling will be handled at the infrastructure level, in the
`forms-deploy` repository.