Move change name form object view partial
Move the view partial for the change name form object to a place in the
views folder which more clearly indicates that it is a partial for a
form object.

This requires overriding `to_partial_path` for the form object class,
and also tweaking the configuration of ActionView to stop it from
namespacing partials.