Fix view tests which show markdown
When running view specs in isolation, for example, the show live view
spec, it fails as GovukFormsMarkdown could not be resolved.

The tests only fail when run on their own.

To fix this issue, the Gemfile is changed to specify the require line.

After making this change, it's possible to remove the individual
require's from other files.