Add method to get Group from Form
We cannot use associations to connect Forms to Group because the
data is spread accross the API and Admin.

To connect Forms to Group, we use the GroupForm model.

It acts a join between the remote ActiveResource based Form and the
local ActiveRecord Group.

This commit adds a new method, group, to the Form model.

This method returns the Group the form belonds to. If the form is not in
a group, it returns nil.