Change Form policy to account for groups
When Form belongs to a group a different
policy applies to #can_view_form?.

Users should be able to view the form if they are a member of the group.
They should not be able to view the form if they are not a member of the
group.

Super_admins are able to view the form if they are in the group or not.