Add account_name
When a user signs in to the service, we have no way of knowing their
name.

As an initial approach we are asking the user enter their name
and organisation when they first use the service.

This commit adds a new subresource under account to show a form to the
user to ask for their name.

At the moment the name form and orgnanisation form are not shown to the
user or linked within the application.

This will change in future commits when the user will be redirected in a
before_action if they do not have name or orgnanisation attributes set.