Use email address as link text on list of users
Previously on the users page, we linked from the user's name to the edit
user page. This failed guideline 2.4.4 of the Web Content Accessibility
Guidelines, as it meant that when multiple users have the same name
(either because their actual names are identical, or because they didn't
have names set), there would be multiple links with the same link
context.

This commit makes the user's email address the link instead of the
user's name. This prevents the issue, as email addresses are unique in
our service.