Add before_action to ask users name & organisation
This commit adds a new before_action to the application controller which
redirects the user to fill out organisation and name if not currently
set.

There are some places we don't want this behaviour. It's been disabled in
the controllers for errors, authentication and health-check.

It's also skipped in the account name and organisation controllers where
it would cause an infinite loop.

The redirects are triggered with separate if tests. If the user has an
organisation but no name, they are redirected to the name page and the
other way round.

This ensures the user sets these values before they use the rest of the
application.