Add integration specs for organisation autocomplete
I want to be able to test how the organisation autocomplete works as a
user would see it.

However, I couldn't think of a way to write these tests in a BDD way, so
I didn't put them in the `spec/features` folder, where we prefer to do
things that way.