Use email address as link text on list of MOUs
Previously on the MOUs page, we linked from the user's name to the edit
user page. This failed guideline 2.4.4 of the Web Content Accessibility
Guidelines, as it meant that when multiple users have the same name
(either because their actual names are identical, or because they didn't
have names set), there would be multiple links with the same link
context.

This commit makes the user's email address the link instead of the
user's name. This prevents the issue, as email addresses are unique in
our service.

We previously made the same fix on the users page (see
[PR 1021](https://github.com/alphagov/forms-admin/pull/1021)).

As part of this work we've also added tests for the view.