Add AfterSignInPathHelper
Add a new module, AfterSignInPathHelper to hold logic for the next path
after a user signs in.

This module will be included into the controller classes which neeed to
know the next path on sign in. It contains the methods to store location
and retreive it when the omniauth callback is called.

Using a module allows us to keep the logic contained, rather than
putting it in controllers directly.