Make auth controller use AfterSignInPathHelper
The authentication_controller now includes the new module
AfterSignInPathHelper.

This gives access to the methods contained in the module.

The controller now uses after_sign_in_next_path to direct the user to
the correct path when the OmniAuth callback is requested after a user
signs in.

The store_location and stored_location method calls are unchanged but
the definitions move from the controller to the module.