Fix login_spec
This spec fails because the trial user does not have an organisation set
and so is redirected to another page.

To fix this, we change to the editor_user. This does have a name and
organisation set and we now editor users are going to stick around into
the future.