Make account controllers use new helpers
The account controllers now include the AfterSignInPathHelper module.

This gives them access to the after_sign_in_next_path method, which they
use to redirect the user when the name/org is successfully updated.

The feature test has been updated to account for the redirect logic
which sends users to the URL they requested before signing in.

The account* request specs have been updated to check that the user is
redirected to whatever URL the after_sign_in_next_path method returns.

The actual method is tested elsewhere so we don't need to test its
behaviour here, just that it is called. To mock a module in a controller
requires using `allow_any_instance_of` because of the way the tests
instantiate objects. We need to turn off the rubocop linter which checks
for these calls.