Put after_sign_in_next_path in module
Wrap the code to work out the next path after sign in into a module,
which can be included in controllers which use it.

This allows us to contain the stored_location code and the code which
checks the User model for org and name.

The tests have to be modified to account for the new module.