Add is_live method to form model
Method mirrors the is_archived method to return whether the form is in
a live state. Currently, we consume the `has_live_version` property on
the form object returned by the API. However, we intend to retire this
property which predates the `state` property of a form as its name is
confusing now that we have archived states which technically have a
live version that is inactive.