Remove usages of form.has_live_version
Replace usages of the `form.has_live_version` property with calls to
the `form.is_live?` method which uses the form state.

The `has_live_version` property will be removed from the form API
response as this property is confusing now that we have archived forms
which technically have a live "version" (a snapshot of the form stored
in the `made_live_forms` table) that is disabled.