Move make live methods into make live form model
Moves the user_wants_to_make_form_live and make_form_live methods out of
the make_live_controller and the unarchive_controller and into the
make_live_form model.

This means that the controller no longer holds logic in private methods
that ought to be in the model's domain, reduces duplication between the
two controllers, and makes it easier to unit test these methods.