Render separate view for making draft of archived form live
We will have different content for the confirm page for making the
draft of an archived form live. Render a separate view for this -
`make_archived_draft_live.html.erb`. This view currently duplicates
the content for making the draft of a live form live, and the content
will need to be updated when it is ready.

This will render the same content regardless of whether the form is in
`archived` or `archived_with_draft` state. The form will only
transition to `archived_with_draft` state if edits are made before
making the draft live. We have a separate route and controller for
making the form live again without making a draft, so the content for
that is separate.