Fix page title on question text page
Previously this page had the same title as the date settings page
(presumably some copy-pasted content that wasn't replaced). As well as
being a bug this was a violation of WCAG 2.4.2 Page Titled, because the
page title did not accurately describe the page.

This commit makes the title match the h1/form field text.