Add new subresource for members of a group
To show the list of users within a group as shown in the designs, we
need a new route.

A new subresrouce is created with only the index. It's called
group_members to follow the convention of group_forms.

"Members" is the content designers choice for users belonging to a
group.