Add a link to to list members of a group
Add a link to the Group#show page to the new page to list members.

Use the list component to contain the change group name and group
members links.

Using the govuk-button group would layout the links horizontally on
desktop and stack them on mobile.

Instead, we use a list, which follows the way we do it with the options
when editing pages.