Add translations for page settings summary fields
Adds translations to the visible labels and change links on the page
settings summary component. This makes it easier to manage our content,
and also means that they change links won't go out of sync with the
visible labels again, causing a regression to the WCAG failure we had
previously.

One consequence of this is that the hidden textin the change links will
start with a capital letter (because the translations we're using are in
sentence case). This shouldn't have any impact on voice control users.
The alternatives would be:
- downcasing the translation (which might have consequences for
pronunciation if any of the translations ever contain intialisms)
- having separate translations for the visible label and hidden text
(which would be harder to maintain and increase the risk of a WCAG
regression).