Make answer type change link consistent
Most of the links in the page settings summary take the form
`Change #{key}`. However, the answer type takes the form
`Change #{key} #{value}`. This is inconsistent and may not match users'
expectations. It's not clear what the benefit of having the extra text
is, since it doesn't disambiguate the link (there should only ever be a
single 'Answer type' field on the page).

This commit removes the extra text, so the change link will just say
"Change Answer type".