Use translations consistently in view spec
Most of our view tests were using the translation in the text they were
querying, but the values for the selection settings were still looking
for untranslated strings. This commit updates those tests to look for
the translated string instead, which should make these tests more
resilient to content changes.