Move logic out of view erb file
This view file had a bit more logic in it than ideal, so this commit
moves the logic for showing and hiding different page settings into the
view.rb instead of the view.html.erb file. This should make it easier to
read the html and see what it's doing.