Use less ambiguous title for new condition page
Previously the new condition page had the same title as the first
page in the new condition journey. This is confusing for users and a
violation of WCAG 2.4.2 Page Titled.

This commit adds additional text to the page title and H1, so that the
page is clearly different from the previous page in the journey.