Add before_action to redirect users to org/name
If a user does not have name/org set, they should not be able to access
other parts of the application before completing them.

This commit adds a before_action, triggered on each request to check
these are complete and redirecting the user if not.

This before_action is skipped on pages which don't need it - errors,
auth, ping, healthcheck, the account controllers themselves.

The redirect only happens if current_user is present (the user is logged
in) so routes like the ping or healthcheck controllers used in
deployment shouldn't need to skip them.

They are still skipped because users (developers?) who have signed in could still hit
them.