Remove unused separator argument
We only use the page_title helper's separator argument in one place (the
new MOU page) and this appears to be a mistake. Without the argument we
would've spotted this mistake, as there would've been an error message.
There's no reason to have this argument, since we don't use it and it's
unclear why we'd want to change the title separator on a case-by-case
basis.

This commit removes the separator argument and fixes the issue in the
MOU file.