Extract Mailchimp sync logic into separate class
The Mailchimp rake task is quite complex, so this commit extracts the
logic for getting and updating the Mailchimp mailing list data into a
separate class. This means that the logic left in the task itself just
deals with finding the correct user data in the database and then
passing it to the new class.

This should make it easier to update the task to deal with both the list
of active users and the list of MOU signers.