Stop duplicates causing flaky tests
Our Faker calls were generating occasional duplicate emails. This meant
that our tests would fail intermittently (because the duplicates would
be removed when the email array was converted into a set, resulting in
the number of emails being wrong).

We can assume that both emails in our database and emails in a given
Mailchimp list are unique, so it should be safe to use
`Faker::Internet.unique.email` here to prevent the issue.