Only add MOU signers to MOU mailing list
Previously our Mailchimp sync would get the list of active users in our
database and sync all of the specified lists with these emails. We only
actually want the platform users mailing list to be updated with these
emails. We only want the MOU signers mailing list to be synced with the
subset of active users in our database who have signed an MOU.

This commit updates the Mailchimp rake task and its configuration to
handle the two mailing lists separately, and to pass a different list of
users to each.