Serve static files by default
For production environments we currently need to have
RAILS_SERVE_STATIC_FILES=1 set in the environment, as Vite places the
frontend assets in the public assets and we don't have another way of
serving them. We don't have any plans to use a separate static asset
server, so we don't really need the flexibility of the environment
variable, and removing the need to set one would simplify our app
deployment configuration and local testing.

This commit removes the line querying the environment variable from the
environment config, so now the app will default to serving files from
the public folder.

This matches the default configuration for Rails 7.1 [[1, 2]].

[1]: https://github.com/rails/rails/pull/47137
[2]: https://github.com/rails/rails/blob/6f0d1ad14b92b9f5906e44740fce8b4f1c7075dc/railties/lib/rails/generators/rails/app/templates/config/environments/production.rb.tt#L25-L26