Use task list component from govuk-components
We had our own custom version of the task list, based on an earlier,
experimental version of the design system component.

Since then the Design System has added an official component for this,
and the govuk-components library has implemented this. This commit
switches to the new component, so that our task list matches other task
lists in government services and we don't have to maintain as much code.