Remove custom html and css to match design system
Our task list was implemented based on an early and experimental version
of the task list component. The Design System now provides a stable
version of the component, so this commit updates the markup to match the
markup provided in the [complete multiple tasks](https://design-system.service.gov.uk/patterns/complete-multiple-tasks/)
guidance.

This also allows us to remove a lot of unused CSS, since we're now using
the default GOV.UK Frontend styles.