Update notification banners with tag argument
Adds the new tag argument to our notification banners. This means we can
use the heading helper without generating inaccessible markup.

The key rule is as follows - the `tag` value provided should be a h3 if
the notification banner also has other content. If the notification
banner does not have other content, the tag should be `p` - since the
'heading' isn't semantically a heading (as there isn't any content
under it).