Move upgrading group logic to service
Having the logic to update the group and send emails to group admins
in the ConfirmUpgradeForm form object confuses the purpose of this
object, which is primarily to deserialize and validate user input form
the page to confirm upgrading a group.

Move this logic to a GroupService. It is anticipated that this might
help with sharing code around email sending when we implement
declining upgrade requests and other group actions.